Move CassandraConstants.getStandardCfDef to ColumnFamilyDefinitions
Shouldn’t be a breaking change as the method was only visible from the
package.
