+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/SchemaMutationLockTables.java
@@ -70,7 +70,7 @@ public class SchemaMutationLockTables {
     }
 
     private void createTableInternal(Cassandra.Client client, TableReference tableRef) throws TException {
-        CfDef cf = CassandraConstants.getStandardCfDef(
+        CfDef cf = ColumnFamilyDefinitions.getStandardCfDef(
                 config.keyspace(),
                 CassandraKeyValueService.internalTableName(tableRef));
         client.system_add_column_family(cf);
